requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
distro==1.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.7.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.15
langchain-openai==0.2.6
langsmith==0.1.140
openai==1.54.2
orjson==3.10.11
packaging==24.1
pdf2image==1.17.0
pillow==11.0.0
pybase64==1.4.0
pydantic==2.9.2
pydantic_core==2.23.4
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
